// hr-frontend/src/lib/api/generated/model/organizationTree.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { OrganizationUnitTreeNode } from './organizationUnitTreeNode';

export interface OrganizationTree {
  organizationId?: number;
  code?: string;
  nameEn?: string;
  nameAr?: string;
  units?: OrganizationUnitTreeNode[];
}