// hr-frontend/src/lib/api/generated/model/organizationUnitTreeNode.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface OrganizationUnitTreeNode {
  id?: number;
  code?: string;
  name?: string;
  nameAr?: string;
  type?: string;
  status?: string;
  children?: OrganizationUnitTreeNode[];
}